Adding unnamed clipboard support
Changing AWTClipboardRegister so that it stops trimming lines.
Before this change when using the AWTClipboardRegister as the default register,
and you yy'ed a line and pasted it, the leading whitespace would be removed. Also
fixed AWTClipboardRegister so that when payload type is lines it does not add an
extra newline to the end which would have doubled up the endin newline when using it as a
clipboard to copy and paste lines.  Adjusted CommandLineParser so that setting
'set clipboard unnamed' in the .vrapperrc file will allow you to use the system
clipboard to copy and paste to and from.