Adding unnamed clipboard support
Changing SWTClipboardRegister so that it stops adding a extra newline to
the system clipboard.  Before this change when using the SWTClipboardRegister
as the default register, and you yy'ed a line and pasted it to another application,
an extra newline would be added to the end.  Adjusted CommandLineParser and added
a configuration listener so that setting 'set clipboard=unnamed' in the .vrapperrc
file will allow you to use the system clipboard to copy and paste to and from.