Close #155 ("add :tabrewind and :tablast").
Note that this uses the "opened order". Eclipse 3.x showed its track in MRU order,
yet the API tracks which file was opened first. This might be confusing at first,
but if no new files are opened or closed, it will always jump to the same files.

If you don't like the MRU behavior, simply switch to the latest Eclipse Juno.