Fix #126 : Implicit dependency on org.eclipse.runtime.compatibility.
The Vrapper plugin was using an old version of the startup extension point where
this startup code could be hosted in the plugin activator. Since Eclipse 3.0,
startup code using the workbench should be moved to a separate class which can
be created later in a different thread without creating two plugin instances.

This commit also fixes a "workbench not created exception" by moving code from
the activator to the startup class - the latter is guaranteed to be run after
workbench creation.