Work on #218: replace status line with SWT widget.
The :tabs command needs to write its multi-line output to the command line buffer,
but this wasn't so much a buffer as it was custom drawing code rendering a string.

Since it's no use re-inventing line wrapping, the status line will now use a
StyledText widget to render its contents. As a bonus, this should allow us to add
a context menu with Copy / Paste / Select all functionality.