Work on #218: Implement UI code, refactor existing code to use it.
Renamed createCommandLine(..) to getCommandLine(..).

Removed "buffer" and "position" fields from AbstractCommandParser,
operate directly on CommandLineUI instead.

EclipseCommandLineUI is now a separate class implementing all the UI
logic. It is master of the current command line contents and selection.
It will also be responsible for drawing the register mode marker.

Refactored PasteRegisterMode to be a mode of its own directly using the
CommandLineUI instead of a hack based on AbstactCommandLineMode.

Refactored SearchCommandParser to grab search contents from
CommandLineUI instead of parser.