Improvements for #218: fix backspace cursor, fix delete previous word.
Upon backspacing the last character, the caret was moved before  the text
content was changed. The caret was thus rendered as a line when it
should be a rectangle.

Delete previous word deleted a whole word plus the letter of the next
word, reverted code to state before command line changes.