Implemented global position marks (A-Z and 0-9).
If capital or digit mark name is used with `m{A-Z0-9}` the mark is
considered global in VIM. So that goto mark operation will switch the
current buffer if necessary. Current implementation uses Eclipse
bookmarks for this feature which make global marks visible, manageable
and persistent across restarts.

In addition to existing goto mark operations `\{a-zA-Z0-9}` was added to
switch to the editor associated with the mark without changing current
position. Thus allowing for editor bookmarks, not position bookmark.
**WARNING**: might conflict with user mappings of `\`.