Fix #212 Cannot execute macro with German keyboard
The @ symbol on a German keyboard is pressed using AltGr + Q. Vrapper
picks up the Alt key, as if Alt + @ is pressed and fails to find the
"playback macro" binding. This might be specific to Windows.

This fix simply binds the playback command to Alt + @ as well as @.
Users should still be able to map anything to Alt + @ if they want.