Start on #251: implement vS and vgS surround keys
* Surround (and optionally indent) using 'S' in visual mode.
Pressing 'S' in visual mode will allow you to press one of the known
delimiters of the regular surround.vim set. These delimiters will then
be used to wrap the current selection.

Linewise mode will put the delimiters on a new line before and after
the selection, then indent the selected text.

Blockmode basically surrounds every line in the block with the
delimitter.

* Surround but don't indent using 'gS' in visual mode.
Pressing 'gS' in visual mode does all of the above except that it won't
indent the selected text.