Alter #251: rewrite TextOperation as a command.
Code was originally written as a TextOperation implementation so that
it could be passed to SelectionBasedTextOperationCommand. Comments on
the issue and the previous commit of igrekster have shown that this way
introduces errors for "surround block" mode.

This commit merges the `VisualCommand` class and
`AddDelimiterToSelectionOperation` so that the code stays in one spot.