Keep dynamic delimiter similar to surround.vim 2.0
Surround 2.0 has no repeat for the "surround selection" modes but it
does have one for the regular commands. If the regular commands are
repeated, the script won't ask again for a replacement delimiter. In
version 1.90, it always asked again.

The code for dynamic delimiters was still based on v1.90, now Vrapper
will remember the last-used delimiter.