Alter #257: do not let cursor go past line ending
When you click in the editor past the end of a line, the cursor is
positioned right at the end of the line - on character further than
vrapper normally alows.

This causes the same exception as before, check if we are at this
point and simply move to the next line.