Retry a key without modifiers if AltGr suspected.
Issue #247 and #212 report that Windows users cannot invoke certain
keybinds if AltGr is needed to press them. For example, Windows reports
AltGr + Q as Ctrl + Alt + @ to vrapper.

Retry any character keys which don't trigger a binding or remap, but
without the control and alt bits set. If that still won't work, it's
likely that no binds were registered for that key.