Reset command line to default mode before paste
Fixes #279 - the change in a2578f64cda4e0f6e4c5d91a764710418f00861f
would reset the command line to default mode after CmdLineUI.type(..)
was called to insert the contents of the register. This corrupts the
command line, as the `"` character is no longer at the expected position.

Change the type(..) method so that it will always reset to DEFAULT mode,
clearing the register marker in any case.