Global map handles default bindings for any mode.
A number of keys in Vim act like equivalents for other keys. Vrapper
had a few of these hard-coded in its input listener, but previous
commits removed this.

This commit re-introduces / adds the following keys which can be used
in any mode:

* CTRL-H: backspace

* CTRL-I: tab

* CTRL-J: return

* CTRL-M: return

* CTRL-[: escape

Because they are in a global map, they don't need to be added to the
initial binding of a mode. However, because remaps are checked first,
users can still remap them should they want to.