Make sure Ctrl+H backspaces in Insert mode.
Fixes #301. Vrapper would remap Ctrl+H to backspace by changing the
input KeyStroke, yet it wouldn't handle the backspace itself and expect
Eclipse to deal with it. Eclipse sent Ctrl+H so it doesn't what to do
and ignores it.

Wrap keystroke as a virtual key to signal that Vrapper should handle it.