Make sure to shift all lines of selection.
A special rule is present to exclude the last line if it is whitespace,
but that rule only checked the selected part. If the user selected only
the indentation of the last line, this check would fail and the last
line wouldn't be indented, even though there's text in it.

Fixes #346.