Make all command definitions local.
The way eclipseaction works is pushing commands into a static variable,
thus creating commands for all tabs at the same time.

Make the CommandLineMode hold both the static coreCommands mapping and
a local mapping for user commands, then pass a combined mapping to the
CommandLineParser.

This alters the way CommandLineParser is constructed, so change all
tests as well.