Fix NullPointerException when sourcing vrapperrc.
During vrapperrc loading, there were 2 CommandLineParser instances in
play. One of the two had a "mapping" field set to null, causing the
exception.

It turns out the second parser is unnecessary, and commands don't need
to be pushed to the CommandLineParser anyway. The CommandLineMode
actually shares its EvaluatorMapping instance with the parser, thus
both are updated at the same time.