Keep cursor near selection end on visual mode exit
Vim will keep the cursor in the very same spot when exiting visual mode
whereas Vrapper would always reset the selection to the top- and
left-most point of the selection.

Try to reset the pointer to a point nearest the end of the selection,
accounting for reverse selections.

Note that this isn't 100% visually correct for visual line-wise mode.
Vim draws a caret inside the selection, whereas Eclipse can only draw
the caret at the end of the selection. Since we now jump to the
position within the selection (as Vim would do), the user might be
slightly surprised to see his caret move. A hack to show the caret in
the "right" position might be added later.

Fixes #391.