Unnamed or named registers shouldn't use clipboard
Vim keeps its unnamed register around even when 'clipboard' is set to
'unnamed', it's just that it is no longer the default.

It is also kept around for when using named registers, as yanking to e.g.
"a will actually update "" in Vim while p will then paste the system
clipboard. Using either ""p or "ap the yanked content will be returned.

This commit makes Vrapper behave in the same way.