Revert "Implement "go to percent of file" motion."
This reverts commit ed69ae8742a45027ec0c5626db1e75687c7c6591.

The border policy will cause the entire line to be erased when trying to
do a d% motion on a pair of parentheses.

On the other hand, switching to INCLUSIVE border policy will cause
runtime errors when running d100% (delete to end of file) due to a wrong
addModelOffset(1) in MotionTextObject.applyBorderPolicy.