Fix count before motion, especially after operator
Add test for a command with 3 count values, one before the register, one
after a register and one before a motion.

ConstructorWrappers.operatorPendingState would wrap all operator
commands in another CountingState. The latter would grab the count value
and pass this to a command and hence override the previous count values.
The third count value would never be passed to the TextObject in
question.

Finally, MultipliedTextObject had a bug where it would simply pass the
requested count value instead of looking at its own count and
multiplying the result.