Place cursor only if no selection, extract code.
While the previous attempt at fixing #211 placed the cursor correctly
for normal mode, it did break reverse mouse selections by chopping off
one character at the end.

The code is extracted to a new class which now listens for mouse clicks.
If the selection is still empty after the user releases the mouse (i.e.
she didn't drag), it is assumed that a click was meant and thus the
caret is placed in the right position.

Work for #440.