Convert keymap name return values to KeyMapInfo
Added a new type KeyMapInfo which holds the active key map name and
the current group name (operator, count, register, ...) which triggered
the active key map.

Thanks to this KeyMapResolver can detect if a keystroke is allowed to
trigger a mapping or not when we are in a dynamic group of which the
length is not known at compile time. Take for example counts where we
don't know if the user will type 1, 2 or more digits; the first digit
could also be a trigger for a mapping but the second digit should be
passed straight to the active EditorMode.

Work for #452.