Allow macro or mapping containing <TAB> key.
The tab is inserted straight to Eclipse which will then place a literal
tab or the equivalent number of spaces according to the Eclipse settings
of the current editor.

This does mean the Vrapper tabstop options are ignored, but at the same
time the user will see the same kind of tab as he would if placing it in
insert.

Fixes #454.