New commands for 'o' / 'O' with correct repetition
Repeating these commands or using these with a count introduces some
interesting behavior: a whole block of similar lines will be inserted,
either before or after the cursor but the cursor will always be at the
end of the inserted block.

Simply repeating InsertLineCommand would sometimes introduce an empty
line too many or put the cursor in the wrong spot, so a new command with
correct repetition logic is created here.

Fixes #507.