Rework SimpleSelection for inclusive visual
When we have a visual selection, the 'selection' option defines if the
character beneath the caret gets included.

We use a hack for left-to-right selections where we draw the caret one
character to the left (see EvilCaret) because Eclipse doesn't support a
Vim-like selection. This means though that the caret can conceptually be
in another spot than just the end of the selection, so make sure to hold
the correct caret position in the SimpleSelection itself.