Fix visual inclusive selection, break visual tests
For starters, the selection behaved like an exclusive one to most of the
codebase, with no proper support for the 'from' and 'to' properties. In
pull request #236 a fix was made to return this caret position, but it
was incomplete as "exclusive" selections actually need different
behavior. The first part is now in properly passing the 'from' and 'to'
fields every time a new visual selection is made. The 'to' field should
now always reflect the position on which the caret is drawn.

Secondly, the code which updated a selection after a motion was
incorrect: reversing the selection one character at a time would
sometimes fail - in such a case the selection would go from +2 to -2.
The code is a lot saner now: it starts with 'from' / 'to' and then makes
sure these two positions are included in the selection. This does mean
that the selection is now really inclusive: the position where a motion
points to gets included.

This is also an alternative fix for #510 as the changes in #236 didn't
take care of Windows newlines, which might cause a Position in the
middle of a Windows newline combination to be returned as the 'to'
position.

Most tests are broken by this commit because the tests don't yet include
the first character on which a motion lands. This is for a following
commit to fix.

Fixes #540.