Calculate correct caret offset; Exception handling
Instead of failing and retrying, just make sure the caret position is
always before or after a newline by checking the line information. This
preservers the position for longer inserts.

In hindsight this commit might not improve much - we only use
smartInsert(String) for single-character inserts and hence the fallback
which does originalOffset + 1 might also work. In the general case, this
workaround might have bitten anybody calling smartInsert with a longer
string.