Port safeAddModelOffset to CursorService, rename.
The previous version always had to scan the current and next line in an
attempt to avoid exceptions. Eclipse doesn't really mind querying for
newlines and can actually immediately tell use where the next line
starts, so the new version should be faster as well as integrated in the
CursorService.

Motions can use it, but they should be aware that the end of file
boundary is silently ignored.