Rework status widget to use styling, be smaller.
Use a custom widget extending from CLabel to allow styling to kick in
while keeping the ability to do custom text shortening (Eclipse by
default replaces the middle of the label with an ellipsis).

The old widget was also using a fixed-width layout to stop the
"recording" text from being shortened / hidden, but this has the result
that other status items might be squeezed. We now use 3 widgets: a
seperator, the mode widget and the "recording" widget, of which the last
one is not always added to the status bar.

Fixes #499, alternative fix for #322 and possible workaround for #357.