Add support for 'clipboard=unnamedplus'
Since the Linux version now supports the selection clipboard, the
"unnamed" variant of the 'clipboard' option *will* copy to the selection
clipboard whereas it used to copy to the system clipboard.

Vim on the other hand always distinguished the clipboards by having an
"unnamedplus" setting, so this commit adds this option to be compatible
with Vim.

Users who currently have `set clipboard=unnamed` are thus advised to
change it to `set clipboard=unnamedplus`.

Windows or Mac users should see no change though.