Always clear selection after visual command.
If you keep the '< and '> mark in the command line window, the command
will be treated like an Ex-range command, in which case the Ex-parser
will set and clear a selection for that range.

If the command line doesn't contain an ex-range, Vrapper needs to make
sure the selection gets cleared after executing the command, otherwise
we go back to normal mode with a selection active.

Fixes #554.