Use block cursor to signal EOL selection.
Before, the selection would always jump to the next line to signal the
the newline character(s) were selected because that is how Eclipse would
do it. Vim on the other hand just selects an extra (whitespace)
character to signal it, so we now shorten the selection and show the
caret on the previous line.

Fixes #561.