Fix offsets and caret for last character selection
If the caret is past the last char instead of 'on' that char, moving the
selection back means you have to press <Left> twice. Simply forbid
moving the caret past that last character when extending the selection.

The cursor can be past the last character if the last line is empty
(in that case it's a newline) and the selection is started from there.
Vim doesn't allow creating an empty selection for that last line though,
so change the selection to be "on" the previous newline. Also make sure
to fix the caret type if this happens.

Related to #540.