'buffers' and 'ls' use Vrapper impl, add 'files'
Whereas `:buffers` and `:ls` used to just list the currently opened
files in an Eclipse window, these commands now call a Vrapper command
which lists all editors in a message window (variant of the command
line).

Since some users might have liked the old behaviour, add a `:files`
command which does the same thing as `:buffers` did previously. This
command exists in Vim as well, so technically it's still a net gain.