Add BufferAndTabService getters, integrate impl
This whole new service needs to have its instance passed down all the
way down.

While it is now a singleton because the InputInterceptorManager
singleton also creates just a single instance, some provisions are made
to deal with window-specific buffer numbers by injecting the
BufferIdManager and passing the BufferAndTabService implementation down
to the interceptor factory.

Integrate so that the currently active editor gets set in this service.