Make sure to match editor on ID and input.
The two-argument function only matches input, which might cause trouble
for multi-page editors. Even if a given input is open in anthor editor,
having that same input in another child page of the current editor will
switch to that page instead of the other editor.

Since we do our best to keep plugin IDs, make sure to use it in the
search.