Filter buffer list to those seen in current window
There might be slight one-offs when unloaded tabs are present in both
windows: if such a tab gets opened in one of the windows, the buffers
which replace the reference in the buffers list might not link to both
windows and hence might not be visible from both.

Switching manually to these kind of editors in both windows should fix
this though.