Activate existing editors directly
Asking the Workbench to open an existing editor or asking a multi-page
editor to activate a given page would trigger some kind of event on that
editor. The XML editor in particular would handle this event by
resetting the cursor, which is jarring to say the least.

Now we check if we can find the editor and then check if the needed page
is active. This deals with the most common usecase where a user would
quickly swap between the current and previous buffer.

Work for #468.