Store last SearchResult, use it to fix end offsets
Previously using a "jump to end of match" search flag would put the
caret on the wrong position: only the length of the search expression
was used. Since regular expressions can expand to longer or shorter
spans of text, we now use the output of the search operation for jumping
to the offset.

Note that still a minor issue remains: jumping to the end offset in
visual-exclusive mode acts like it is inclusive, Vim on the other hand
does treat it as exclusive (this conflicts with the documentation where
it says the 'e' flag should make it inclusive, but YMMV).

Fixes #539.