Implement `gn` / `gN` for visual mode.
These bindings will now do a search with a forced direction and offset,
jumping to the end of the next match or the beginning of the previous
match.

The behaviour of `gn` and `gN` isn't 100% correct while in Linewise
Visual mode or Blockwise visual mode (it should change the mode to
visual and act as if gn was ran there). This small annoyance is hard to
fix without duplicating visual mode code, and since the user is likely
to start off in visual mode after pressing 'gn' / 'gN', this is left
unsolved until somebody specially asks for it.

Work for #291.