Rework replace method to skip using dynamic offsets
When using regex find/replace, the length of the keyword or replacement
string are irrelevant.

Eclipse allows us to get information of the length of a match and the
length of the replacement string. We now use that information that know
how many characters to skip and how a replace affects the line length.

Fixes #570.