Don't close command line widget if reopened.
The close() call would hide the widget asynchronously as a fix for #508.
This assumption is no longer valid if we switch to message mode because
then we get an open() call in the same UI job, yet the "hide" UI job is
still waiting to be executed.

Use an extra status flag to know if the hide code should run.