Allow 'let' for the search register, highlighting
Using `let @/=''` now clears the last search and resets search
highlighting.

Passing a keyword also works, this will update the search highlighting
for the given keyword. The search direction and other settings will be
used from the last search, or if the editor is fresh then defaults are
used.

Fixes #577.