Make sure blockwise change command edits all lines
The ChangeOperation would try to detect if the invoking mode is
blockwise visual by checking if a selection was passed in and the last
selection was blockwise.

After 22025355db79f8ef2785402846087cf2457d7f this detection got broken,
but it's actually far simpler to remove this detection and make
blockwise visual mode call the right operation in the first place.

Fixes #604.