Apply blockwise indent in middle of line, not start
Selecting a block and inserting indentation in Vim actually inserts
whitespace at or after the left block edge, not at the beginning of the
line.

Similarly, selecting a block and removing indent should only operate on
whitespace after the left edge of the block selection.

This code was not explicitly written in our blockwise mode so the
standard InsertShiftWidth operation was invoked from AbstractVisualMode.

Work for #612.