Make 'source ~/somefile' resolve to $HOME/somefile
Vim has a concept of a current directory, sourcing a relative path will
hence resolve according to that directory. It is thus common to prepend
'~/somefile' to mean a file in the home directory without needing to
give an absolute path.

Vrapper on the other hand would resolve all relative paths to the home
directory by default. For Vim compatibility the '~' at the beginning is
ignored, otherwise it would look for $HOME/~/somefile.