Rename Selection.selectMarks to reset, bounds check.
"reset" is a better name now that both Position parameters are from/to.
Introduce a parent class to bound check these parameters (for inclusive
as well as exclusive mode).

Let DefaultAdapter call this method so that bounds are properly reset,
there was a chance for #376 to occur if the left boundary was > than the
length of the document.