Fix searching backward for a single char
Eclipse always sticks to its current position when searching for a
single char. It is therefore important to move the current position back
or forwards depending on where we want to look for the next match.

Longer keywords don't need this "hack" so the old code made this move
conditional, even though it is easier and just as good to alter the
position in all cases.

Fixes #605.