Allow disable of local configuration listeners.
The local configuration listeners were originally made for the purpose
of reacting to settings changes while Vrapper is active, e.g. changing
the highlighting when `set ignorecase` is invoked after highlighting is
already active. They might fire too soon though as reading .vrapperrc
will also change the in-memory configuration.

That's why this commit adds a way to toggle the firing of these
listeners.