Explicitly activate highlight on init, not config.
When the "hlscope" setting is set to "window", the highlight from
previous editors also needs to be present upon initialization of a fresh
editor.

The local config listeners in SearchMode.java are triggered by loading
.vrapperrc but they are not useful here, the fresh editor is not yet in
the buffer list so running that code will only reset the highlighting in
previous editors (even clearing it when "clear" is set). Disable those
local config listeners and simply activate highlighting for the current
editor.

If we use any other hlscope setting we simply do nothing, which is what
we want: nothing gets cleared, nothing gets added.

Extra fix for #577.