Allow block mode visualisation to be backwards.
Some of the code introduced when "toEOL" was added should only be active
when calculating the bounds for an entire region, since the "normal
block" selection actually used to work fine.

This code would interfere with selections where the "to" offset was more
to the left than the "from" offset, the block would then be drawn as
just a single character wide even though it stretched to the left.