Serialize keystroke with modifiers as <C-A-S-x>
Recording a macro will need to store keypresses like 'Ctrl-R' to a
register, so the key is converted to a string code which the keycode
parser understands when reading the register for execution.

Before 51deea3b49, all Ctrl-something keystrokes were internally using
some special key character: Ctrl-A is 0x01, Ctrl-B is 0x02 and so on.
These non-printable characters were then saved in the registers, similar
to how Vim does.

After the mentioned commit, KeyStrokes internally have the character
they're from plus modifier flags yet the conversion mechanism wasn't
changed. This resulted in 'Ctrl-R' being stored as just 'R'.

This commit changes the code to take the modifier flags and write the
keystroke in the "special key" format, e.g. <C-R>.

Fixes #628.