Reset Vrapper selection for all selection changes.
StyledText does not call the SelectionListener for all of its selection
changes as some JFace commands managed to change the selection without
triggering an event.

This meant that triggering an Eclipse selection command while in Visual
mode would show a different selection in the UI than what Vrapper would
use when executing a ChangeOperation.

It should now be legal to do this in visual mode and it should work.
While it's still not recommended for visual line mode, that shouldn't
"glitch" either.