Store position before and after jump in Eclipse.
The Eclipse navigation history is used for our "jump list"
implementation. However, we didn't really tell Eclipse when we were
making a jump nor telling it where the cursor was.

A few new methods in CursorService should now allow Eclipse to grab the
current position and store it in its navigation history.

Fixes #487.