:set with empty string clears string-set option
When calling :set clipboard=autoselect and then :set clipboard=, the
option would actually not be cleared. This is because the empty String
was directly passed to the validate function which in turn could not
find an empty string in the list of allowed values.

Simply check for an empty string and clear all values.