Implement debug logging (conditional output to info)
This mechanism should be used sparingly as all the code still gets
executed in the background. If any call to debug() remain, it should
only execute once per editor.

It is however nice to be able to switch logging on or off without
removing swathes of code in between test runs.