Make sure to update edit marks after cleanindent.
When creating a new line with `o`, the startposition is recorded right
after creating the new line and its indent.

This indent gets cleaned up when the user quits Insert mode without
having inserted anything. The startposition wouldn't be modified though,
meaning that it points to an offset which is now outside the new line.
Not only are the edit marks wrong, a user reported a problem where this
wrong offset might go outside the bounds of the file, triggering
exceptions when registering the edit marks.

Simply recalculate the offsets to compensate for the removed indent.

Fixes #652.