EXPERIMENTAL Force default caret visible, move it
Basically we always force the caret to be visible (block select mode
be damned) and then shift the caret boundaries to where we want it to
be.

So far this seems to work on Windows and at is technically native. While
it is an ugly hack to change the bounds of the caret, it is far easier to
implement than the hacks we would need to do for a blinking caret with
custom painting code.

This code still has a problem with placing the caret behind the line,
but so did the previous attempt.