Refactor safeAddModelOffset, add shiftPositionForViewOffset
Shuffling the delta and targetOffset around allows us to reuse the same
function both for view and model offsets. The trick for view offsets
lies in calculating the targetOffset before converting to model offsets,
this way +/-1 can jump over a fold and hence cover a large span in model
space offsets.