Handle inclusive selection workaround in view space.
Calculating the "to" member of Selection should happen in view space,as
calculating it in model coordinates might mean that we move the caret
into a collapsed fold. View offsets are always visible, we only need to
make sure that we aren't in a newline combo (for which we call
shiftPositionForViewOffset).

Make sure not to use SimpleSelection's constructor with a TextRange, that
method will calculate the "to" member using model offsets.

Fixes #657.