Selecting text object should open fold for caret
A fold in e.g. a Java source file might show a line ending with a '{' in
case a method gets folded. Vim only allows to place the caret at the
start of this "fold line" but Vrapper users can navigate to the end and
trigger a "va{" command ("select outer braces") textobject command. The
'to' member of a selection then points to the ending brace which is
hidden by the fold.  Now the caret is no longer hidden when making this
selection.

This fix also resolves a bug where the MoveRightAcrossLines motion jumps
to the start of the file when the caret was hidden: the view offset for
a hidden caret is -1, so incrementing it would jump to offset 0.