EXPERIMENTAL Remove shifted caret hacks from core
The caret painter can actually place the caret easily if we can use
setLocation. The EvilCaret used by LEFT_SHIFTED_RECTANGLE overrides
setLocation though, moving even more to the left.

Therefore this commit removes the use of LEFT_SHIFTED_RECTANGLE in
Vrapper core, core now simply uses a plain RECTANGLE and Vrapper's
Eclipse implementation does the rest. All other caret types create a
plain Eclipse Caret widget instead of the EvilCaret.

The only remaining place to use LEFT_SHIFTED_RECTANGLE is when the user
has their mouse pressed somewhere behind a line and we're still in
normal mode, in such a case we want to make it appear like the caret is
"on" the last character of that line. That code is in Vrapper's Eclipse
plugin though.